Podcast Player: memoize colors object (#15336)
* extract colors object generator

* remove extraneous checks and optimizations
